code review task kontak java part 1

+++ src/main/java/id/solvrtech/kontakjava/App.java
@@ -6,6 +6,7 @@ import id.solvrtech.kontakjava.service.PersonService;
 
 import java.sql.Connection;
 import java.util.ArrayList;
+import java.util.List;
 
 import static id.solvrtech.kontakjava.utils.Helper.*;
 import static id.solvrtech.kontakjava.utils.MysqlConnection.getConnection;
@@ -30,7 +31,6 @@ public class App {
         // create new personService
         PersonService personService = new PersonService();
 
-
         while (true) {
             // call the showMenu method
             showMenu();
@@ -39,20 +39,20 @@ public class App {
 
             if (menuChoice == 1) {
                 System.out.println("=== ALL KONTAK PERSONS ===");
-                ArrayList<Person> persons = personService.getAll();
+                List<Person> persons = personService.getAll();
                 showData(persons);
 
                 pressEnterToContinue();
 
             } else if (menuChoice == 2) {
                 System.out.println("=== SEARCH PERSON ===");
-                String searchInput = readInputAsString("Enter your name or your: ");
+                String searchInput = readInputAsString("Enter your name or your phone number: ");
 
-                ArrayList<Person> searchResult = personService.searchPerson(searchInput);
+                List<Person> searchResult = personService.searchPerson(searchInput);
                 if (searchInput != null) {
                     showData(searchResult);
                     pressEnterToContinue();
-                }else{
+                } else {
                     System.out.println("Search not found!");
                     pressEnterToContinue();
                 }
@@ -60,27 +60,29 @@ public class App {
             } else if (menuChoice == 3) {
                 System.out.println("=== CREATE NEW PERSON ===");
 
-                boolean valid = true;
-                while (valid) {
+                boolean inValid = true;
+                while (inValid) {
                     String name = readInputAsString("Name: ");
 
                     if (!name.isEmpty()) {
-                        while (valid) {
+                        while (inValid) {
                             String phoneNumber = readInputAsString("Phone number: ");
 
-                            if (isPhoneNumberExits(personService.getAll(), phoneNumber) == true) {
-                                System.out.println("Your phone number is already exists, please input another number!");
+                            if (isInputNumeric(phoneNumber)) {
+                                Person newData = personService.create(name, phoneNumber);
 
-                            } else if (isInputNumeric(phoneNumber)) {
-                                personService.create(name, phoneNumber);
+                                if (newData == null) {
+                                    System.out.println("Your phone number is already exists, please input another number!");
 
-                                System.out.println(" ");
-                                System.out.println("New person has been created!");
+                                } else {
 
-                                valid = false;
+                                    System.out.println(" ");
+                                    System.out.println("New person has been created!");
+
+                                    inValid = false;
+                                }
                             } else {
                                 System.out.println("Invalid phone number!");
-                                System.out.println(" ");
                             }
                         }
                     } else {
@@ -93,7 +95,7 @@ public class App {
 
             } else if (menuChoice == 4) {
                 System.out.println("=== EDIT PERSON ===");
-                ArrayList<Person> persons = showData(personService.getAll());
+                List<Person> persons = showData(personService.getAll());
 
                 if (persons.size() != 0) {
                     int id = readInputAsInt("Choose the ID of the person to edit: ");
@@ -108,23 +110,27 @@ public class App {
                         System.out.println(" ");
 
                         System.out.println("=== INPUT NEW DATA ===");
-                        boolean valid = true;
-                        while (valid) {
+                        boolean inValid = true;
+                        while (inValid) {
                             String newName = readInputAsString("Name: ");
 
                             if (!newName.isEmpty()) {
-                                while (valid) {
+                                while (inValid) {
                                     String newPhoneNumber = readInputAsString("Phone number: ");
 
-                                    // validasi jika ada number yang sama dengan data yang lain
-                                    if (isPhoneNumberExits(personService.getAll(), newPhoneNumber, thePerson.getId()) == true) {
-                                        System.out.println("Your phone number is already exists, please input another number!");
-                                    } else if (isInputNumeric(newPhoneNumber)) {
-                                        personService.update(thePerson, newName, newPhoneNumber);
-                                        System.out.println(" ");
-                                        System.out.println("New person has been updated!");
-                                        pressEnterToContinue();
-                                        valid = false;
+                                    if (isInputNumeric(newPhoneNumber)) {
+                                        Person editData = personService.update(thePerson, newName, newPhoneNumber);
+
+                                        if (editData == null) {
+                                            System.out.println("Your phone number is already exists, please input another number!");
+                                        } else {
+                                            System.out.println(" ");
+
+                                            System.out.println("New person has been updated!");
+                                            pressEnterToContinue();
+
+                                            inValid = false;
+                                        }
                                     } else {
                                         System.out.println("Invalid phone number!");
                                     }
@@ -135,43 +141,58 @@ public class App {
                         System.out.println("Person data not found!");
                         pressEnterToContinue();
                     }
-                }else{
+                } else {
                     pressEnterToContinue();
                 }
 
-            }else if (menuChoice == 5) {
+            } else if (menuChoice == 5) {
                 System.out.println("=== DELETE PERSON ===");
-                ArrayList<Person> persons = showData(personService.getAll());
+                List<Person> persons = showData(personService.getAll());
 
                 if (persons.size() != 0) {
                     int id = readInputAsInt("Choose the ID of the person to delete: ");
                     Person thePerson = personService.getById(id);
 
-                    // Tampilkan data yang sudah dipilih
-                    System.out.println("=== Person Data ===");
-                    System.out.println("Nama         : " + thePerson.getName());
-                    System.out.println("Phone number : " + thePerson.getPhoneNumber());
-                    System.out.println(" ");
-                    String confirm = readInputAsString("Are you sure you want to delete this person? (y/n): ");
-
-                    if (confirm.equals("y")) {
-                        String deleted = personService.delete(id);
-                        System.out.println(deleted);
+                    if (thePerson != null) {
+                        // Tampilkan data yang sudah dipilih
+                        System.out.println("=== Person Data ===");
+                        System.out.println("Nama         : " + thePerson.getName());
+                        System.out.println("Phone number : " + thePerson.getPhoneNumber());
                         System.out.println(" ");
-                    } else if (confirm.equals("n")) {
-                        System.out.println("Cancel deletion!");
-                    } else {
-                        System.out.println("Please input 'y' or 'n' to continue!");
-                    }
-                }
 
-            }else if (menuChoice == 6) {
+                        boolean inValid = true;
+                        while (inValid) {
+                            String confirm = readInputAsString("Are you sure you want to delete this person? (y/n): ");
+
+                            if (confirm.equals("y")) {
+                                String deleted = personService.delete(id);
+                                System.out.println(deleted);
+                                System.out.println(" ");
+                                pressEnterToContinue();
+                                inValid = false;
+                            } else if (confirm.equals("n")) {
+                                System.out.println("Cancel deletion!");
+                                pressEnterToContinue();
+                                inValid = false;
+                            } else {
+                                System.out.println("Please input 'y' or 'n' to continue!");
+                                pressEnterToContinue();
+                            }
+                        }
+                    } else {
+                        System.out.println("Person data not found!");
+                        pressEnterToContinue();
+                    }
+                }else{
+                    pressEnterToContinue();
+                }
+            } else if (menuChoice == 6) {
                 System.out.println("=== EXIT ===");
                 System.out.println("See you soon!");
                 break;
-            }else {
-                System.out.println("PLease input the correct option!");
-                pressEnterToContinue();
+            } else {
+                System.out.println("Please input the correct option!");
+                System.out.println(" ");
             }
         }
     }
